requirements.txt
musicbrainzngs==0.7.1
pylistenbrainz==0.5.1
pypresence==4.2.1
python-dotenv==1.0.0
requests==2.28.2
